+++ pygeos/creation.py
@@ -13,6 +13,8 @@ __all__ = [
     "multipolygons",
     "geometrycollections",
     "box",
+    "prepare",
+    "destroy_prepared",
 ]
 
 
@@ -170,3 +172,46 @@ def geometrycollections(geometries):
         An array of geometries
     """
     return lib.create_collection(geometries, GeometryType.GEOMETRYCOLLECTION)
+
+
+def prepare(geometry, **kwargs):
+    """Compute a prepared geometry.
+
+    A prepared geometry is a normal geometry with added information such as an
+    index on the line segments. This improves the performance of the following operations:
+    contains, contains_properly, covered_by, covers, crosses, disjoint, intersects,
+    overlaps, touches, and within.
+
+    Note that if a prepared geometry is modified, the newly created Geometry object is
+    not prepared. In that case, ``prepare`` should be called again.
+
+    This function does not recompute previously prepared geometries; 
+    it is efficient to call this function on an array that partially contains prepared geometries. 
+
+    Parameters
+    ----------
+    geometry : Geometry or array_like
+
+    See also
+    --------
+    destroy_prepared
+    """
+    return lib.prepare(geometry, **kwargs)
+
+
+def destroy_prepared(geometry, **kwargs):
+    """Destroy a previously prepared geometry, freeing up memory.
+
+    Note that the prepared geometry will always be cleaned up if the geometry itself
+    is dereferenced. This function needs only be called in very specific circumstances,
+    such as freeing up memory without losing the geometries, or benchmarking.
+
+    Parameters
+    ----------
+    geometry : Geometry or array_like
+
+    See also
+    --------
+    prepare
+    """
+    return lib.destroy_prepared(geometry, **kwargs)