+++ src/pygeom.h
@@ -7,6 +7,7 @@
 
 typedef struct {
   PyObject_HEAD void* ptr;
+  void* ptr_prepared;
 } GeometryObject;
 
 extern PyTypeObject GeometryType;
@@ -15,6 +16,8 @@ extern PyTypeObject GeometryType;
 extern PyObject* GeometryObject_FromGEOS(GEOSGeometry* ptr, GEOSContextHandle_t ctx);
 /* Get a GEOSGeometry from a GeometryObject */
 extern char get_geom(GeometryObject* obj, GEOSGeometry** out);
+extern char get_geom_with_prepared(GeometryObject* obj, GEOSGeometry** out,
+                                   GEOSPreparedGeometry** prep);
 
 extern int init_geom_type(PyObject* m);