+++ src/ufuncs.c
@@ -196,16 +196,7 @@ static void O_b_func(char** args, npy_intp* dimensions, npy_intp* steps, void* d
 static PyUFuncGenericFunction O_b_funcs[1] = {&O_b_func};
 
 /* Define the geom, geom -> bool functions (YY_b) */
-static void* disjoint_data[1] = {GEOSDisjoint_r};
-static void* touches_data[1] = {GEOSTouches_r};
-static void* intersects_data[1] = {GEOSIntersects_r};
-static void* crosses_data[1] = {GEOSCrosses_r};
-static void* within_data[1] = {GEOSWithin_r};
-static void* contains_data[1] = {GEOSContains_r};
-static void* overlaps_data[1] = {GEOSOverlaps_r};
 static void* equals_data[1] = {GEOSEquals_r};
-static void* covers_data[1] = {GEOSCovers_r};
-static void* covered_by_data[1] = {GEOSCoveredBy_r};
 typedef char FuncGEOS_YY_b(void* context, void* a, void* b);
 static char YY_b_dtypes[3] = {NPY_OBJECT, NPY_OBJECT, NPY_BOOL};
 static void YY_b_func(char** args, npy_intp* dimensions, npy_intp* steps, void* data) {
@@ -244,6 +235,74 @@ finish:
   GEOS_FINISH_THREADS;
 }
 static PyUFuncGenericFunction YY_b_funcs[1] = {&YY_b_func};
+
+/* Define the geom, geom -> bool functions (YY_b) prepared */
+static void* contains_func_tuple[2] = {GEOSContains_r, GEOSPreparedContains_r};
+static void* contains_data[1] = {contains_func_tuple};
+// static void* contains_properly_func_tuple[2] = {..., GEOSPreparedContainsProperly_r};
+// static void* contains_properly_data[1] = {contains_properly_func_tuple};
+static void* covered_by_func_tuple[2] = {GEOSCoveredBy_r, GEOSPreparedCoveredBy_r};
+static void* covered_by_data[1] = {covered_by_func_tuple};
+static void* covers_func_tuple[2] = {GEOSCovers_r, GEOSPreparedCovers_r};
+static void* covers_data[1] = {covers_func_tuple};
+static void* crosses_func_tuple[2] = {GEOSCrosses_r, GEOSPreparedCrosses_r};
+static void* crosses_data[1] = {crosses_func_tuple};
+static void* disjoint_func_tuple[2] = {GEOSDisjoint_r, GEOSPreparedDisjoint_r};
+static void* disjoint_data[1] = {disjoint_func_tuple};
+static void* intersects_func_tuple[2] = {GEOSIntersects_r, GEOSPreparedIntersects_r};
+static void* intersects_data[1] = {intersects_func_tuple};
+static void* overlaps_func_tuple[2] = {GEOSOverlaps_r, GEOSPreparedOverlaps_r};
+static void* overlaps_data[1] = {overlaps_func_tuple};
+static void* touches_func_tuple[2] = {GEOSTouches_r, GEOSPreparedTouches_r};
+static void* touches_data[1] = {touches_func_tuple};
+static void* within_func_tuple[2] = {GEOSWithin_r, GEOSPreparedWithin_r};
+static void* within_data[1] = {within_func_tuple};
+static char YY_b_p_dtypes[3] = {NPY_OBJECT, NPY_OBJECT, NPY_BOOL};
+static void YY_b_p_func(char** args, npy_intp* dimensions, npy_intp* steps, void* data) {
+  FuncGEOS_YY_b* func = ((FuncGEOS_YY_b**)data)[0];
+  FuncGEOS_YY_b* func_prepared = ((FuncGEOS_YY_b**)data)[1];
+
+  GEOSGeometry *in1 = NULL, *in2 = NULL;
+  GEOSPreparedGeometry* in1_prepared = NULL;
+  char ret;
+
+  GEOS_INIT_THREADS;
+
+  BINARY_LOOP {
+    /* get the geometries: return on error */
+    if (!get_geom_with_prepared(*(GeometryObject**)ip1, &in1, &in1_prepared)) {
+      errstate = PGERR_NOT_A_GEOMETRY;
+      goto finish;
+    }
+    if (!get_geom(*(GeometryObject**)ip2, &in2)) {
+      errstate = PGERR_NOT_A_GEOMETRY;
+      goto finish;
+    }
+    if ((in1 == NULL) | (in2 == NULL)) {
+      /* in case of a missing value: return 0 (False) */
+      ret = 0;
+    } else {
+      if (in1_prepared == NULL) {
+        /* call the GEOS function */
+        ret = func(ctx, in1, in2);
+      } else {
+        /* call the prepared GEOS function */
+        ret = func_prepared(ctx, in1_prepared, in2);
+      }
+      /* return for illegal values */
+      if (ret == 2) {
+        errstate = PGERR_GEOS_EXCEPTION;
+        goto finish;
+      }
+    }
+    *(npy_bool*)op1 = ret;
+  }
+
+finish:
+
+  GEOS_FINISH_THREADS;
+}
+static PyUFuncGenericFunction YY_b_p_funcs[1] = {&YY_b_p_func};
 
 /* Define the geom -> geom functions (Y_Y) */
 static void* envelope_data[1] = {GEOSEnvelope_r};
@@ -363,6 +422,55 @@ static void Y_Y_func(char** args, npy_intp* dimensions, npy_intp* steps, void* d
   free(geom_arr);
 }
 static PyUFuncGenericFunction Y_Y_funcs[1] = {&Y_Y_func};
+
+/* Define the geom -> no return value functions (Y) */
+static char PrepareGeometryObject(void* ctx, GeometryObject* geom) {
+  if (geom->ptr_prepared == NULL) {
+    geom->ptr_prepared = (GEOSPreparedGeometry*)GEOSPrepare_r(ctx, geom->ptr);
+    if (geom->ptr_prepared == NULL) {
+      return PGERR_GEOS_EXCEPTION;
+    }
+  }
+  return PGERR_SUCCESS;
+}
+static char DestroyPreparedGeometryObject(void* ctx, GeometryObject* geom) {
+  if (geom->ptr_prepared != NULL) {
+    GEOSPreparedGeom_destroy_r(ctx, geom->ptr_prepared);
+    geom->ptr_prepared = NULL;
+  }
+  return PGERR_SUCCESS;
+}
+
+static void* prepare_data[1] = {PrepareGeometryObject};
+static void* destroy_prepared_data[1] = {DestroyPreparedGeometryObject};
+typedef char FuncPyGEOS_Y(void* ctx, GeometryObject* geom);
+static char Y_dtypes[1] = {NPY_OBJECT};
+static void Y_func(char** args, npy_intp* dimensions, npy_intp* steps, void* data) {
+  FuncPyGEOS_Y* func = (FuncPyGEOS_Y*)data;
+  GEOSGeometry* in1 = NULL;
+  GeometryObject* geom_obj = NULL;
+
+  GEOS_INIT;
+
+  NO_OUTPUT_LOOP {
+    geom_obj = *(GeometryObject**)ip1;
+    if (!get_geom(geom_obj, &in1)) {
+      errstate = PGERR_GEOS_EXCEPTION;
+      goto finish;
+    }
+    if (in1 != NULL) {
+      errstate = func(ctx, geom_obj);
+      if (errstate != PGERR_SUCCESS) {
+        goto finish;
+      }
+    }
+  }
+
+finish:
+
+  GEOS_FINISH;
+}
+static PyUFuncGenericFunction Y_funcs[1] = {&Y_func};
 
 /* Define the geom, double -> geom functions (Yd_Y) */
 static void* GEOSInterpolateProtectEmpty_r(void* context, void* geom, double d) {
@@ -2203,9 +2311,19 @@ TODO relate functions
                                   PyUFunc_None, #NAME, "", 0);                   \
   PyDict_SetItemString(d, #NAME, ufunc)
 
+#define DEFINE_YY_b_p(NAME)                                                          \
+  ufunc = PyUFunc_FromFuncAndData(YY_b_p_funcs, NAME##_data, YY_b_p_dtypes, 1, 2, 1, \
+                                  PyUFunc_None, #NAME, "", 0);                       \
+  PyDict_SetItemString(d, #NAME, ufunc)
+
 #define DEFINE_Y_Y(NAME)                                                       \
   ufunc = PyUFunc_FromFuncAndData(Y_Y_funcs, NAME##_data, Y_Y_dtypes, 1, 1, 1, \
                                   PyUFunc_None, #NAME, "", 0);                 \
+  PyDict_SetItemString(d, #NAME, ufunc)
+
+#define DEFINE_Y(NAME)                                                                   \
+  ufunc = PyUFunc_FromFuncAndData(Y_funcs, NAME##_data, Y_dtypes, 1, 1, 0, PyUFunc_None, \
+                                  #NAME, "", 0);                                         \
   PyDict_SetItemString(d, #NAME, ufunc)
 
 #define DEFINE_Yi_Y(NAME)                                                        \
@@ -2274,16 +2392,16 @@ int init_ufuncs(PyObject* m, PyObject* d) {
   DEFINE_O_b(is_missing);
   DEFINE_O_b(is_valid_input);
 
-  DEFINE_YY_b(disjoint);
-  DEFINE_YY_b(touches);
-  DEFINE_YY_b(intersects);
-  DEFINE_YY_b(crosses);
-  DEFINE_YY_b(within);
-  DEFINE_YY_b(contains);
-  DEFINE_YY_b(overlaps);
+  DEFINE_YY_b_p(disjoint);
+  DEFINE_YY_b_p(touches);
+  DEFINE_YY_b_p(intersects);
+  DEFINE_YY_b_p(crosses);
+  DEFINE_YY_b_p(within);
+  DEFINE_YY_b_p(contains);
+  DEFINE_YY_b_p(overlaps);
   DEFINE_YY_b(equals);
-  DEFINE_YY_b(covers);
-  DEFINE_YY_b(covered_by);
+  DEFINE_YY_b_p(covers);
+  DEFINE_YY_b_p(covered_by);
 
   DEFINE_Y_Y(envelope);
   DEFINE_Y_Y(convex_hull);
@@ -2295,6 +2413,9 @@ int init_ufuncs(PyObject* m, PyObject* d) {
   DEFINE_Y_Y(extract_unique_points);
   DEFINE_Y_Y(get_exterior_ring);
   DEFINE_Y_Y(normalize);
+
+  DEFINE_Y(prepare);
+  DEFINE_Y(destroy_prepared);
 
   DEFINE_Yi_Y(get_point);
   DEFINE_Yi_Y(get_interior_ring);