+++ CHANGELOG.rst
@@ -27,6 +27,9 @@ Version 0.9 (unreleased)
   ``get_num_geometries``, ``get_num_interior_rings``, ``get_num_points``) now return 0
   for ``None`` input values instead of -1 (#218).
 * Fixed internal GEOS error code detection for ``get_dimensions`` and ``get_srid`` (#218).
+* Addition of ``prepare`` function that generates a GEOS prepared geometry which is stored on
+  the Geometry object itself. All binary predicates (except ``equals``) make use of this (#92).
+
 
 Version 0.8 (2020-09-06)
 ------------------------